[Chore] abstract out a new file `editorCommand.ts`
